Enable lint related Rubocop rules (#9231)
* Add Rubocop Config to temporary disable cops

* Fix Lint/RedundantSafeNavigation  rubocop

* Fix Lint/RedundantCopEnableDirective rubocop

* Fix Layout/DefEndAlignment rubocop

* Fix Lint/UnmodifiedReduceAccumulator

* Fix Lint/DuplicateBranch

* Fix Lint/SymbolConversion

* Fixing Lint/EmptyClass

* Fix Lint/AmbiguousOperatorPrecedence

* Revert .. Lint/UnmodifiedReduceAccumulator

* Ignore Lint/UnmodifiedReduceAccumulator

* Fix: Lint/NoReturnInBeginEndBlocks

* Fix Lint/EmptyBlock

* Fix failing spec

* Cleaning .rubocop-disabled.yml

* Implement review Recommendations

* Refactor changelog_generator

* Fix Lint/UnmodifiedReduceAccumulator

+++ decidim-core/app/commands/decidim/gallery_methods.rb
@@ -59,13 +59,9 @@ module Decidim
 
     def photo_cleanup!
       gallery_attached_to.photos.each do |photo|
-        if @form.photos.map(&:id).exclude?(photo.id)
-          if @form.respond_to?(:documents) && @form.documents.map(&:id).exclude?(photo.id)
-            photo.destroy!
-          elsif !@form.respond_to?(:documents)
-            photo.destroy!
-          end
-        end
+        next unless @form.photos.map(&:id).exclude?(photo.id)
+
+        photo.destroy! if (@form.respond_to?(:documents) && @form.documents.map(&:id).exclude?(photo.id)) || !@form.respond_to?(:documents)
       end
       # manually reset cached photos
       gallery_attached_to.reload
